split Featured into container/component
